增加系统管理相关的表结构设计 重新生成系统管理相关表、服务、mapper

+++ src/main/java/com/github/jamsa/rap/controller/UserController.java
@@ -15,12 +15,12 @@ import java.util.Map;
  */
 @Controller
 @RequestMapping("/user")
-public class UserController extends BaseEntityController<User,Long> {
+public class UserController extends BaseEntityController<User,Integer> {
     @Autowired
     UserService userService;
 
     @Override
-    public BaseEntityService<User, Long> getService() {
+    public BaseEntityService<User, Integer> getService() {
         return userService;
     }
 }
